In rate in float

+++ shared/types/api/node.go
@@ -443,16 +443,16 @@ type GetSdStatusResponse struct {
 }
 
 type SdStatusResponse struct {
-	SdUtilizerLatestBalance   *big.Int `json:"sdUtilizerLatestBalance"`
-	SdUtilizedBalance         *big.Int `json:"sdUtilizedBalance"`
-	SdCollateralCurrentAmount *big.Int `json:"sdCollateralCurrentAmount"`
-	SdCollateralRequireAmount *big.Int `json:"sdCollateralRequireAmount"`
-	SdMaxUtilizableAmount     *big.Int `json:"sdMaxUtilizableAmount"`
-	UtilizationRate           *big.Int `json:"utilizationRate"`
-	SdBalance                 *big.Int `json:"sdBalance"`
-	PoolAvailableSDBalance    *big.Int `json:"poolAvailableSDBalance"`
-	SdRewardEligible          *big.Int `json:"sdRewardEligible"`
-	NotEnoughSdCollateral     bool     `json:"notEnoughSdCollateral"`
+	SdUtilizerLatestBalance   *big.Int   `json:"sdUtilizerLatestBalance"`
+	SdUtilizedBalance         *big.Int   `json:"sdUtilizedBalance"`
+	SdCollateralCurrentAmount *big.Int   `json:"sdCollateralCurrentAmount"`
+	SdCollateralRequireAmount *big.Int   `json:"sdCollateralRequireAmount"`
+	SdMaxUtilizableAmount     *big.Int   `json:"sdMaxUtilizableAmount"`
+	UtilizationRate           *big.Float `json:"utilizationRate"`
+	SdBalance                 *big.Int   `json:"sdBalance"`
+	PoolAvailableSDBalance    *big.Int   `json:"poolAvailableSDBalance"`
+	SdRewardEligible          *big.Int   `json:"sdRewardEligible"`
+	NotEnoughSdCollateral     bool       `json:"notEnoughSdCollateral"`
 }
 
 type NodeRepayExcessSDResponse struct {
